对toggle组件进行化简
