+++ src/components/shadcn/toggle/style.css
@@ -2,6 +2,8 @@
     background-color: hsl(var(--accent));
 }
 
+
+
 .data-\[state\=pressed\]\:bg-primary[pressed] {
     background-color: hsl(var(--accent));
 }
@@ -13,5 +15,11 @@
 
 .rounded-sm {
     border-radius: 4px;
+}
+
+
+/* 让 toggle 组件内 svg 在 hover 时变为 50% 透明度 */
+[data-hover='true'] svg {
+    opacity: 0.5;
 }
