+++ src/components/shadcn/toggle/toggle.ts
@@ -7,14 +7,9 @@ import {
     ShadcnToggleContextType,
 } from './interface';
 import { optimizeTailwindClasses } from '@/www/utils/tailwind';
-import {ShadcnToggleContent} from './toggle-content'
 import { ToggleExposes } from '@/core/behaviors/as-toggle/interface';
 
-if (!customElements.get('shadcn-toggle-content')) {
-    customElements.define('shadcn-toggle-content', ShadcnToggleContent);
-}
-
-// shadcn/ui toggle 基础样式
+// shadcn/ui newyork 风格 toggle 基础样式
 const baseClasses = [
     'w-9',
     'h-9',
@@ -30,14 +25,19 @@ const baseClasses = [
     'focus-visible:ring-offset-2',
     'disabled:opacity-50',
     'disabled:pointer-events-none',
-    
+    'cursor-pointer',
+    'bg-background',
 ];
 
-// 状态样式
 const stateClasses = [
-    'cursor-pointer',
     'data-[state=pressed]:bg-primary',
     'data-[state=pressed]:text-primary-foreground',
+];
+
+const iconClasses = [
+    'h-4',
+    'w-4',
+    'pointer-events-none',
 ];
 
 const SHADCN_TOGGLE_DEFAULT_CLASS = [...baseClasses, ...stateClasses].join(' ');
@@ -48,37 +48,43 @@ export const ShadcnTogglePrototype = definePrototype<ShadcnToggleProps, ToggleEx
         // 注入 Toggle 行为
         asToggle(p);
 
-        p.context.provide(ShadcnToggleContext, (updateContext) => {
-            const context: ShadcnToggleContextType = {
-                toggleRef: null,
-                updateRef: (name, ref) => {
-                    const originalRef = context[name];
-                    updateContext({
-                        [name]: ref,
-                    });
-                    if (originalRef && originalRef !== ref) originalRef.remove();
-                },
-            };
-            return context;
-        });
-
+        
         let _originalCls = '';
 
         p.lifecycle.onMounted(() => {
             _originalCls = p.view.getElement().className;
-
-            const hostElement = p.view.getElement();
-
-            if (!hostElement.querySelector('shadcn-toggle-content')) {
-                const toggleContent = document.createElement('shadcn-toggle-content');
-                hostElement.appendChild(toggleContent);
-            }
         });
 
-        return () => {
+        // 渲染函数
+        return (renderer) => {
             const hostElement = p.view.getElement();
             const allClasses = [SHADCN_TOGGLE_DEFAULT_CLASS, _originalCls].filter(Boolean).join(' ');
             hostElement.className = optimizeTailwindClasses(allClasses);
+
+            // 清空已有内容，防止重复渲染
+            hostElement.innerHTML = '';
+
+            // SVG 图标，状态与 toggle 保持一致
+            const h = renderer.createElement;
+            const svg = h(
+                'svg',
+                {
+                    viewBox: '0 0 24 24',
+                    fill: 'none',
+                    stroke: 'currentColor',
+                    strokeWidth: '2',
+                    strokeLinecap: 'round',
+                    strokeLinejoin: 'round',
+                    class: iconClasses.join(' '),
+                },
+                [
+                    h('path', { d: 'M6 4h8a4 4 0 0 1 0 8H6z' }),
+                    h('path', { d: 'M6 12h9a4 4 0 0 1 0 8H6z' }),
+                ]
+            );
+            hostElement.appendChild(svg);
+
+
         };
     },
 });
