+++ src/core/behaviors/as-toggle/index.ts
@@ -1,5 +1,5 @@
 export { default as asToggle } from './as-toggle';
-export { default as asToggleContent } from './as-toggle-content';
+
 
 export type { ToggleProps, ToggleState, ToggleContextType, ToggleExposes } from './interface';
